Code review improvements: UB, size_t, swap, memory checks, radix sort opt

+++ include/sortings/heap_sort.h
@@ -3,8 +3,8 @@
 
 #include <stddef.h>
 
-void bottom_up_heapify(int arr[], int n, int i, int k);
-void build_k_heap     (int arr[], int n, int k);
-void heap_sort_k      (int arr[], int n, int k);
+void bottom_up_heapify(int arr[], size_t n, size_t i, int k);
+void build_k_heap     (int arr[], size_t n, int k);
+void heap_sort_k      (int arr[], size_t n, int k);
 
 #endif // HEAP_SORT_H
